media: xilinx: vip: Add xvip_device_ops to prepare for s_stream rework
To prepare for the move to the V4L2 subdev .enable_streams() and
.disable_streams() pad operations, introduce a xvip_device_ops structure
with corresponding .enable_streams() and .disable_streams() operations.

Three new helpers are also added: xvip_s_stream(), xvip_enable_streams()
and xvip_disable_streams() are drop-in implementation of the
corresponding subdev operations. They are implemented to replicate the
existing behaviour of individual drivers and of the Xilinx V4L2
framework. This will allow refactoring the framework in reviewable steps
without modifying all drivers at the same time.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>