media: i2c: max9286: Fix routing table parsing in link configuration
When iterating over the routing table, the route index is used to check
if the route is included in the route_mask, while the route sink pad
should be used instead as route_mask stores a bitmask of sink pads.
Fixing this means that we don't need the loop index anymore, so we can
switch to using the for_each_active_route() macro is inactive routes can
be ignored. Given that the route_mask check was meant to skip inactive
routes, it can also be dropped as a result, simplifying the code.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>