media: xilinx: dma: Store number of inputs and outputs separately
The xvip_pipeline structure stores the total number of DMA engines in
the pipeline, as well as the number of engines that have been started.
That information is used to handle pipeline start and stop. Split it
into separate counters for the input and output DMA engines, to enable
more complex decisions to be made at start and stop time.

As the amount of code in xvip_pipeline_set_stream() becomes larger as a
result, and because none is shared between the start and stop paths,
split the function in two, xvip_pipeline_start_dma() and
xvip_pipeline_stop_dma().

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>